// src/resources/Purpose_Plan/Purpose.interface.ts
export interface Ipurpose {
  purposeName: string;
  projectName: string;
  whatIsIt: string;
  why: string;
  whyYou: string;
  theVision: string;
  theReach: string;
  theMission: string;
  whatDoesItDo: string;
  whomDoesItServe: string;
  whyShouldItExist: string;
  defineYourPurpose: string;
  steps: string;
  values: string;
  personalities: string
  userId?: any;
}